+++ print_pointer.c
@@ -6,18 +6,19 @@
 /*   By: syeresko <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/24 20:18:34 by syeresko          #+#    #+#             */
-/*   Updated: 2018/11/24 20:56:43 by syeresko         ###   ########.fr       */
+/*   Updated: 2018/11/25 20:16:59 by syeresko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include "ft_printf.h"
+#include <inttypes.h>
 
 #include <unistd.h>		// for write in ft_print_pointer
 
 
 
-static char	*ft_pointer_s(unsigned long long num)
+static char	*ft_pointer_s(uintmax_t num)
 {
 	char	*s;
 	char	digit;
@@ -53,7 +54,7 @@ int		ft_print_pointer(const t_fmt *f, void *ptr)
 	int		n_digits;
 	int		val_len;
 
-	s = ft_pointer_s((unsigned long long)ptr);
+	s = ft_pointer_s((uintmax_t)ptr);
 	n_prefix = 2;
 	n_digits = ft_strlen(s);
 	n_zeroes = count_zeroes_x(f, n_prefix, n_digits);
