+++ print_unsigned.c
@@ -6,12 +6,13 @@
 /*   By: syeresko <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/23 17:05:38 by syeresko          #+#    #+#             */
-/*   Updated: 2018/11/24 13:33:28 by syeresko         ###   ########.fr       */
+/*   Updated: 2018/11/25 20:18:00 by syeresko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include "ft_printf.h"
+#include <inttypes.h>
 
 #include <unistd.h>		// for write in ft_print_integer
 
@@ -48,7 +49,7 @@ static char	*ft_unsigned_s(const t_fmt *a_fmt, unsigned long long num)
 }
 */
 
-static char	*ft_unsigned_s(unsigned long long num)
+static char	*ft_unsigned_s(uintmax_t num)
 {
 	char	*s;
 
@@ -102,7 +103,7 @@ int			ft_print_unsigned(const t_fmt *a_fmt, unsigned long long num)
 }
 */
 
-int			ft_print_unsigned(const t_fmt *f, unsigned long long num)
+int			ft_print_unsigned(const t_fmt *f, uintmax_t num)
 {
 	char	*s;
 	int		n_zeroes;
